+++ src/jfan/fan/sys/Test.java
@@ -81,7 +81,7 @@ public class Test
     }
     if (expected != null && actual != null)
     {
-      if (hash(expected).longValue() != hash(actual).longValue())
+      if (hash(expected) != hash(actual))
       {
         fail("Equal but different hash codes: " +
           expected + " (0x" + FanInt.toHex(hash(expected)) + ") ?= " +
